+++ src/host/rpc_connections.h
@@ -50,10 +50,10 @@ namespace asynchost
   class ConnIDGenerator
   {
   public:
-    /// This is the same as tls::ConnID and quic::ConnID
+    /// This is the same as ccf::tls::ConnID and quic::ConnID
     using ConnID = int64_t;
-    static_assert(std::is_same<tls::ConnID, quic::ConnID>());
-    static_assert(std::is_same<tls::ConnID, ConnID>());
+    static_assert(std::is_same<::tls::ConnID, quic::ConnID>());
+    static_assert(std::is_same<::tls::ConnID, ConnID>());
 
     ConnIDGenerator() : next_id(1) {}
 
@@ -121,7 +121,7 @@ namespace asynchost
         LOG_DEBUG_FMT("rpc read {}: {}", id, len);
 
         RINGBUFFER_WRITE_MESSAGE(
-          tls::tls_inbound,
+          ::tls::tls_inbound,
           parent.to_enclave,
           id,
           serializer::ByteRange{data, len});
@@ -139,7 +139,7 @@ namespace asynchost
       {
         if constexpr (isTCP<ConnType>())
         {
-          RINGBUFFER_WRITE_MESSAGE(tls::tls_close, parent.to_enclave, id);
+          RINGBUFFER_WRITE_MESSAGE(::tls::tls_close, parent.to_enclave, id);
         }
       }
     };
@@ -185,7 +185,7 @@ namespace asynchost
         if constexpr (isTCP<ConnType>())
         {
           RINGBUFFER_WRITE_MESSAGE(
-            tls::tls_start, parent.to_enclave, peer_id, interface_name);
+            ::tls::tls_start, parent.to_enclave, peer_id, interface_name);
           return;
         }
 
@@ -328,7 +328,7 @@ namespace asynchost
       // Invalidating the TCP socket will result in the handle being closed. No
       // more messages will be read from or written to the TCP socket.
       sockets[id] = nullptr;
-      RINGBUFFER_WRITE_MESSAGE(tls::tls_close, to_enclave, id);
+      RINGBUFFER_WRITE_MESSAGE(::tls::tls_close, to_enclave, id);
 
       return true;
     }
@@ -348,9 +348,9 @@ namespace asynchost
       messaging::Dispatcher<ringbuffer::Message>& disp)
     {
       DISPATCHER_SET_MESSAGE_HANDLER(
-        disp, tls::tls_outbound, [this](const uint8_t* data, size_t size) {
+        disp, ::tls::tls_outbound, [this](const uint8_t* data, size_t size) {
           auto [id, body] =
-            ringbuffer::read_message<tls::tls_outbound>(data, size);
+            ringbuffer::read_message<::tls::tls_outbound>(data, size);
 
           ConnID connect_id = (ConnID)id;
           LOG_DEBUG_FMT("rpc write from enclave {}: {}", connect_id, body.size);
@@ -359,9 +359,9 @@ namespace asynchost
         });
 
       DISPATCHER_SET_MESSAGE_HANDLER(
-        disp, tls::tls_connect, [this](const uint8_t* data, size_t size) {
+        disp, ::tls::tls_connect, [this](const uint8_t* data, size_t size) {
           auto [id, host, port] =
-            ringbuffer::read_message<tls::tls_connect>(data, size);
+            ringbuffer::read_message<::tls::tls_connect>(data, size);
 
           LOG_DEBUG_FMT("rpc connect request from enclave {}", id);
 
@@ -377,16 +377,17 @@ namespace asynchost
         });
 
       DISPATCHER_SET_MESSAGE_HANDLER(
-        disp, tls::tls_stop, [this](const uint8_t* data, size_t size) {
-          auto [id, msg] = ringbuffer::read_message<tls::tls_stop>(data, size);
+        disp, ::tls::tls_stop, [this](const uint8_t* data, size_t size) {
+          auto [id, msg] =
+            ringbuffer::read_message<::tls::tls_stop>(data, size);
 
           LOG_DEBUG_FMT("rpc stop from enclave {}, {}", id, msg);
           stop(id);
         });
 
       DISPATCHER_SET_MESSAGE_HANDLER(
-        disp, tls::tls_closed, [this](const uint8_t* data, size_t size) {
-          auto [id] = ringbuffer::read_message<tls::tls_closed>(data, size);
+        disp, ::tls::tls_closed, [this](const uint8_t* data, size_t size) {
+          auto [id] = ringbuffer::read_message<::tls::tls_closed>(data, size);
 
           LOG_DEBUG_FMT("rpc closed from enclave {}", id);
           close(id);