+++ src/node/rpc/frontend.h
@@ -13,6 +13,7 @@
 #include "ccf/service/tables/nodes.h"
 #include "ccf/service/tables/service.h"
 #include "common/configuration.h"
+#include "enclave/enclave_time.h"
 #include "enclave/rpc_handler.h"
 #include "forwarder.h"
 #include "http/http_jwt.h"
@@ -68,22 +69,6 @@ namespace ccf
     {
       history = tables.get_history().get();
       endpoints.set_history(history);
-    }
-
-    void update_metrics(
-      const std::shared_ptr<ccf::RpcContextImpl>& ctx,
-      const endpoints::EndpointDefinitionPtr& endpoint)
-    {
-      int cat = ctx->get_response_status() / 100;
-      switch (cat)
-      {
-        case 4:
-          endpoints.increment_metrics_errors(endpoint);
-          return;
-        case 5:
-          endpoints.increment_metrics_failures(endpoint);
-          return;
-      }
     }
 
     endpoints::EndpointDefinitionPtr find_endpoint(
@@ -442,7 +427,6 @@ namespace ccf
           ccf::errors::InvalidAuthenticationInfo,
           "Invalid authentication credentials.",
           std::move(json_details));
-        update_metrics(ctx, endpoint);
       }
 
       return identity;
@@ -482,7 +466,7 @@ namespace ccf
           HTTP_STATUS_NOT_IMPLEMENTED,
           ccf::errors::NotImplemented,
           "Request cannot be forwarded to primary on HTTP/2 interface.");
-        update_metrics(ctx, endpoint);
+
         return;
       }
 
@@ -492,7 +476,7 @@ namespace ccf
           HTTP_STATUS_INTERNAL_SERVER_ERROR,
           ccf::errors::InternalError,
           "No consensus or forwarder to forward request.");
-        update_metrics(ctx, endpoint);
+
         return;
       }
 
@@ -504,7 +488,7 @@ namespace ccf
           HTTP_STATUS_SERVICE_UNAVAILABLE,
           ccf::errors::RequestAlreadyForwarded,
           "RPC was already forwarded.");
-        update_metrics(ctx, endpoint);
+
         return;
       }
 
@@ -522,7 +506,7 @@ namespace ccf
           HTTP_STATUS_SERVICE_UNAVAILABLE,
           ccf::errors::InternalError,
           "RPC could not be forwarded to unknown primary.");
-        update_metrics(ctx, endpoint);
+
         return;
       }
 
@@ -536,7 +520,7 @@ namespace ccf
           HTTP_STATUS_SERVICE_UNAVAILABLE,
           ccf::errors::InternalError,
           "Unable to establish channel to forward to primary.");
-        update_metrics(ctx, endpoint);
+
         return;
       }
 
@@ -555,9 +539,46 @@ namespace ccf
     void process_command(std::shared_ptr<ccf::RpcContextImpl> ctx)
     {
       size_t attempts = 0;
-      constexpr auto max_attempts = 30;
       endpoints::EndpointDefinitionPtr endpoint = nullptr;
 
+      const auto start_time = ccf::get_enclave_time();
+
+      process_command_inner(ctx, endpoint, attempts);
+
+      const auto end_time = ccf::get_enclave_time();
+
+      if (endpoint != nullptr)
+      {
+        endpoints::RequestCompletedEvent rce;
+        rce.method = endpoint->dispatch.verb.c_str();
+        rce.dispatch_path = endpoint->dispatch.uri_path;
+        rce.status = ctx->get_response_status();
+        // Although enclave time returns a microsecond value, the actual
+        // precision/granularity depends on the host's TimeUpdater. By default
+        // this only advances each millisecond. Avoid implying more precision
+        // than that, by rounding to milliseconds
+        rce.exec_time = std::chrono::duration_cast<std::chrono::milliseconds>(
+          end_time - start_time);
+        rce.attempts = attempts;
+
+        endpoints.handle_event_request_completed(rce);
+      }
+      else
+      {
+        endpoints::DispatchFailedEvent dfe;
+        dfe.method = ctx->get_method();
+        dfe.status = ctx->get_response_status();
+
+        endpoints.handle_event_dispatch_failed(dfe);
+      }
+    }
+
+    void process_command_inner(
+      std::shared_ptr<ccf::RpcContextImpl> ctx,
+      endpoints::EndpointDefinitionPtr& endpoint,
+      size_t& attempts)
+    {
+      constexpr auto max_attempts = 30;
       while (attempts < max_attempts)
       {
         if (consensus != nullptr)
@@ -600,18 +621,6 @@ namespace ccf
         if (endpoint == nullptr)
         {
           return;
-        }
-        else
-        {
-          // Only register calls to existing endpoints
-          if (attempts == 1)
-          {
-            endpoints.increment_metrics_calls(endpoint);
-          }
-          else
-          {
-            endpoints.increment_metrics_retries(endpoint);
-          }
         }
 
         try
@@ -701,7 +710,6 @@ namespace ccf
 
           if (!ctx->should_apply_writes())
           {
-            update_metrics(ctx, endpoint);
             return;
           }
 
@@ -778,7 +786,6 @@ namespace ccf
                 history->try_emit_signature();
               }
 
-              update_metrics(ctx, endpoint);
               return;
             }
 
@@ -794,7 +801,7 @@ namespace ccf
                 HTTP_STATUS_SERVICE_UNAVAILABLE,
                 ccf::errors::TransactionReplicationFailed,
                 "Transaction failed to replicate.");
-              update_metrics(ctx, endpoint);
+
               return;
             }
           }
@@ -811,15 +818,15 @@ namespace ccf
         {
           ctx->clear_response_headers();
           ctx->set_error(std::move(e.error));
-          update_metrics(ctx, endpoint);
+
           return;
         }
-        catch (const JsonParseError& e)
+        catch (const ccf::JsonParseError& e)
         {
           ctx->clear_response_headers();
           ctx->set_error(
             HTTP_STATUS_BAD_REQUEST, ccf::errors::InvalidInput, e.describe());
-          update_metrics(ctx, endpoint);
+
           return;
         }
         catch (const nlohmann::json::exception& e)
@@ -827,7 +834,7 @@ namespace ccf
           ctx->clear_response_headers();
           ctx->set_error(
             HTTP_STATUS_BAD_REQUEST, ccf::errors::InvalidInput, e.what());
-          update_metrics(ctx, endpoint);
+
           return;
         }
         catch (const kv::KvSerialiserException& e)
@@ -846,7 +853,7 @@ namespace ccf
             HTTP_STATUS_INTERNAL_SERVER_ERROR,
             ccf::errors::InternalError,
             e.what());
-          update_metrics(ctx, endpoint);
+
           return;
         }
       } // end of while loop
@@ -859,7 +866,7 @@ namespace ccf
           "Transaction continued to conflict after {} attempts. Retry "
           "later.",
           max_attempts));
-      update_metrics(ctx, endpoint);
+
       static constexpr size_t retry_after_seconds = 3;
       ctx->set_response_header(http::headers::RETRY_AFTER, retry_after_seconds);